Changed structure to LinkedHashMap to retain mapping order
Thanks @maurerit. The HashMap does not retain order on k/v pairs and, as such, the first key retrieval that's performed for filtering purposes would fail if the first key didn't happen to be alphabetically first. My naming scheme for testing was a coincidence that prevented discovery of this nuance. The LinkedHashMap does retain order, and has solved the problem.